Fix signedness comparison in chat length check
GCC's -Wsign-compare (on the Linux CI) flagged comparing buf.size()
(std::string::size_type, unsigned) against iMaxChatMessageLength (int).
Cast the option to size_t;
the length is never negative, so the result is unchanged.

+++ src/server/CServer_Parse.cpp
@@ -441,7 +441,7 @@ void CServerNetEngine::ParseChatText(CBytestream *bs) {
 	//Resize oversized messages and optionally kick the sender
 	//TODO: Implement check at lower level to drop packets without ever passing them to chat handler? Or would this cause problems or confusion?
 	//TODO: Limit message length client side too?
-	if (tLXOptions->bCheckChatMessageLength && buf.size() > tLXOptions->iMaxChatMessageLength){
+	if (tLXOptions->bCheckChatMessageLength && buf.size() > (size_t)tLXOptions->iMaxChatMessageLength){
 		buf.resize(tLXOptions->iMaxChatMessageLength);
 		if (tLXOptions->bKickOversizedMsgSenders){
 			server->DropClient(cl, CLL_KICK, "Attempt to send oversized message or command");
